+++ ext/standard/basic_functions.c
@@ -5838,7 +5838,7 @@ static void php_ini_parser_cb_with_sections(zval *arg1, zval *arg2, zval *arg3, 
 PHP_FUNCTION(parse_ini_file)
 {
 	char *filename = NULL;
-	int filename_len = 0;
+	size_t filename_len = 0;
 	zend_bool process_sections = 0;
 	zend_long scanner_mode = ZEND_INI_SCANNER_NORMAL;
 	zend_file_handle fh;
@@ -5879,7 +5879,7 @@ PHP_FUNCTION(parse_ini_file)
 PHP_FUNCTION(parse_ini_string)
 {
 	char *string = NULL, *str = NULL;
-	int str_len = 0;
+	size_t str_len = 0;
 	zend_bool process_sections = 0;
 	zend_long scanner_mode = ZEND_INI_SCANNER_NORMAL;
 	zend_ini_parser_cb_t ini_parser_cb;
