+++ ext/standard/http.c
@@ -239,7 +239,7 @@ PHP_FUNCTION(http_build_query)
 {
 	zval *formdata;
 	char *prefix = NULL, *arg_sep=NULL;
-	int arg_sep_len = 0, prefix_len = 0;
+	size_t arg_sep_len = 0, prefix_len = 0;
 	smart_str formstr = {0};
 	zend_long enc_type = PHP_QUERY_RFC1738;
